feature/era-glonass-integration-r2: VIN synchronizer:
moved `VinSynchronizerMonad` to `Types` sub-module
